requirements.txt
streamlit
google-generativeai
python-dotenv

#pip install -r requirements.txt